@@ -25,68 +25,68 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.2.12",
-    "@angular/cdk": "^18.2.13",
-    "@angular/common": "^18.2.12",
-    "@angular/compiler": "^18.2.12",
-    "@angular/core": "^18.2.12",
-    "@angular/forms": "^18.2.12",
-    "@angular/material": "^18.2.13",
-    "@angular/platform-browser": "^18.2.12",
-    "@angular/platform-browser-dynamic": "^18.2.12",
-    "@angular/router": "^18.2.12",
+    "@angular/animations": "^19.0.0",
+    "@angular/cdk": "^19.0.0",
+    "@angular/common": "^19.0.0",
+    "@angular/compiler": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/forms": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@angular/platform-browser": "^19.0.0",
+    "@angular/platform-browser-dynamic": "^19.0.0",
+    "@angular/router": "^19.0.0",
     "@auth0/angular-jwt": "^5.2.0",
     "@cisstech/nge": "^18.0.4",
-    "@myrmidon/auth-jwt-admin": "5.1.0",
-    "@myrmidon/auth-jwt-login": "5.1.0",
-    "@myrmidon/cadmus-mat-physical-size": "^5.1.4",
-    "@myrmidon/cadmus-mat-physical-state": "^0.0.3",
-    "@myrmidon/cadmus-refs-asserted-chronotope": "5.1.3",
-    "@myrmidon/cadmus-refs-asserted-ids": "^5.1.7",
-    "@myrmidon/cadmus-refs-assertion": "5.1.1",
-    "@myrmidon/cadmus-refs-decorated-counts": "^5.1.1",
-    "@myrmidon/cadmus-refs-doc-references": "^5.1.1",
-    "@myrmidon/cadmus-refs-geonames-lookup": "^5.1.1",
-    "@myrmidon/cadmus-refs-historical-date": "5.1.2",
-    "@myrmidon/cadmus-refs-lookup": "^5.1.2",
-    "@myrmidon/cadmus-refs-proper-name": "5.1.4",
-    "@myrmidon/cadmus-refs-viaf-lookup": "^5.1.0",
-    "@myrmidon/cadmus-text-block-view": "^5.1.0",
-    "@myrmidon/cadmus-text-ed": "^5.1.0",
-    "@myrmidon/cadmus-text-ed-md": "^5.1.2",
-    "@myrmidon/cadmus-text-ed-txt": "^5.1.0",
-    "@myrmidon/ng-mat-tools": "3.1.0",
-    "@myrmidon/ng-tools": "3.1.1",
-    "@myrmidon/paged-data-browsers": "^2.0.5",
-    "@ngneat/effects-ng": "^3.1.4",
+    "@myrmidon/auth-jwt-admin": "6.0.0",
+    "@myrmidon/auth-jwt-login": "6.0.0",
+    "@myrmidon/cadmus-mat-physical-size": "^6.0.0",
+    "@myrmidon/cadmus-mat-physical-state": "^1.0.0",
+    "@myrmidon/cadmus-refs-asserted-chronotope": "6.0.0",
+    "@myrmidon/cadmus-refs-asserted-ids": "^6.0.0",
+    "@myrmidon/cadmus-refs-assertion": "6.0.0",
+    "@myrmidon/cadmus-refs-decorated-counts": "^6.0.0",
+    "@myrmidon/cadmus-refs-doc-references": "^6.0.0",
+    "@myrmidon/cadmus-refs-geonames-lookup": "^6.0.0",
+    "@myrmidon/cadmus-refs-historical-date": "6.0.0",
+    "@myrmidon/cadmus-refs-lookup": "^6.0.0",
+    "@myrmidon/cadmus-refs-proper-name": "6.0.0",
+    "@myrmidon/cadmus-refs-viaf-lookup": "^6.0.0",
+    "@myrmidon/cadmus-text-block-view": "^6.0.0",
+    "@myrmidon/cadmus-text-ed": "^6.0.0",
+    "@myrmidon/cadmus-text-ed-md": "^6.0.0",
+    "@myrmidon/cadmus-text-ed-txt": "^6.0.0",
+    "@myrmidon/ng-mat-tools": "4.0.0",
+    "@myrmidon/ng-tools": "4.0.0",
+    "@myrmidon/paged-data-browsers": "^3.0.0",
     "@swimlane/ngx-graph": "^8.4.0",
     "@types/diff-match-patch": "^1.0.36",
     "diff-match-patch": "^1.0.5",
     "fast-deep-equal": "^3.1.3",
-    "marked": "^14.1.3",
+    "marked": "^15.0.2",
     "monaco-editor": "^0.52.0",
     "rxjs": "~7.8.1",
     "ts-md5": "^1.3.1",
-    "tslib": "^2.8.0",
-    "zone.js": "~0.14.10"
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.12",
-    "@angular/cli": "^18.2.12",
-    "@angular/compiler-cli": "^18.2.12",
-    "@angular/localize": "^18.2.12",
+    "@angular-devkit/build-angular": "^19.0.1",
+    "@angular/cli": "^19.0.1",
+    "@angular/compiler-cli": "^19.0.0",
+    "@angular/localize": "^19.0.0",
     "@types/d3-dispatch": "^3.0.6",
     "@types/d3-drag": "^3.0.7",
+    "@types/d3-scale": "^4.0.8",
     "@types/d3-timer": "^3.0.2",
     "@types/jasmine": "~5.1.4",
-    "@types/node": "^22.8.4",
+    "@types/node": "^22.9.1",
     "jasmine-core": "~5.4.0",
     "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^18.2.1",
-    "typescript": "~5.5.3"
+    "ng-packagr": "^19.0.1",
+    "typescript": "~5.6.2"
   }
 }
