@@ -1,6 +1,6 @@
 {
   "name": "@myrmidon/cadmus-flags-ui",
-  "version": "8.1.3",
+  "version": "9.0.0",
   "description": "Cadmus - item flags UI.",
   "keywords": [
     "Cadmus"
@@ -14,13 +14,13 @@
     "name": "Daniele Fusi"
   },
   "dependencies": {
-    "tslib": "^2.5.0"
+    "tslib": "^2.3.0"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@myrmidon/ng-tools": "^3.1.1",
-    "@myrmidon/ng-mat-tools": "^3.1.0"
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@myrmidon/ng-tools": "^4.0.0",
+    "@myrmidon/ng-mat-tools": "^4.0.0"
   },
   "sideEffects": false
 }
