@@ -1,6 +1,6 @@
 {
   "name": "@myrmidon/cadmus-graph-pg",
-  "version": "8.1.1",
+  "version": "9.0.0",
   "description": "Cadmus - semantic graph page.",
   "keywords": [
     "Cadmus"
@@ -14,14 +14,14 @@
     "name": "Daniele Fusi"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@myrmidon/ng-tools": "^3.0.0",
-    "@myrmidon/ng-mat-tools": "^3.1.0",
-    "@myrmidon/cadmus-api": "^8.1.0",
-    "@myrmidon/cadmus-graph-ui": "^8.1.1"
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@myrmidon/ng-tools": "^4.0.0",
+    "@myrmidon/ng-mat-tools": "^4.0.0",
+    "@myrmidon/cadmus-api": "^9.0.0",
+    "@myrmidon/cadmus-graph-ui": "^9.0.0"
   },
   "dependencies": {
-    "tslib": "^2.5.0"
+    "tslib": "^2.3.0"
   }
 }
