@@ -1,6 +1,6 @@
 {
   "name": "@myrmidon/cadmus-graph-ui-ex",
-  "version": "8.1.2",
+  "version": "9.0.0",
   "description": "Cadmus - extensions to semantic graph components.",
   "keywords": [
     "Cadmus"
@@ -14,18 +14,18 @@
     "name": "Daniele Fusi"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@swimlane/ngx-graph": "^8.3.0",
-    "@myrmidon/ng-tools": "^3.1.1",
-    "@myrmidon/ng-mat-tools": "^3.1.0",
-    "@myrmidon/cadmus-refs-lookup": "^5.1.2",
-    "@myrmidon/cadmus-core": "^8.1.0",
-    "@myrmidon/cadmus-api": "^8.1.0",
-    "@myrmidon/cadmus-graph-ui": "^8.1.1"
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@swimlane/ngx-graph": "^8.4.0",
+    "@myrmidon/ng-tools": "^4.0.0",
+    "@myrmidon/ng-mat-tools": "^4.0.0",
+    "@myrmidon/cadmus-refs-lookup": "^6.0.0",
+    "@myrmidon/cadmus-core": "^9.0.0",
+    "@myrmidon/cadmus-api": "^9.0.0",
+    "@myrmidon/cadmus-graph-ui": "^9.0.0"
   },
   "dependencies": {
-    "tslib": "^2.5.0"
+    "tslib": "^2.3.0"
   },
   "sideEffects": false
 }
