@@ -1,6 +1,6 @@
 {
   "name": "@myrmidon/cadmus-graph-ui",
-  "version": "8.1.3",
+  "version": "9.0.0",
   "description": "Cadmus - semantic graph components.",
   "keywords": [
     "Cadmus"
@@ -14,16 +14,16 @@
     "name": "Daniele Fusi"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@angular/material": "^18.0.0",
-    "@myrmidon/ng-tools": "^3.1.1",
-    "@myrmidon/ng-mat-tools": "^3.1.0",
-    "@myrmidon/cadmus-api": "^8.1.0",
-    "@myrmidon/paged-data-browsers": "^1.1.0",
-    "@myrmidon/cadmus-refs-lookup": "^5.1.2"
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@angular/material": "^19.0.0",
+    "@myrmidon/ng-tools": "^4.0.0",
+    "@myrmidon/ng-mat-tools": "^4.0.0",
+    "@myrmidon/cadmus-api": "^9.0.0",
+    "@myrmidon/paged-data-browsers": "^3.0.0",
+    "@myrmidon/cadmus-refs-lookup": "^6.0.0"
   },
   "dependencies": {
-    "tslib": "^2.5.0"
+    "tslib": "^2.3.0"
   }
 }
