@@ -1,6 +1,6 @@
 {
   "name": "@myrmidon/cadmus-item-editor",
-  "version": "8.1.5",
+  "version": "9.0.0",
   "description": "Cadmus - item editor.",
   "keywords": [
     "Cadmus"
@@ -14,15 +14,15 @@
     "name": "Daniele Fusi"
   },
   "peerDependencies": {
-    "@angular/common": "^18.0.0",
-    "@angular/core": "^18.0.0",
-    "@myrmidon/ng-tools": "^3.1.1",
-    "@myrmidon/ng-mat-tools": "^3.1.0",
-    "@myrmidon/paged-data-browsers": "^2.0.0",
-    "@myrmidon/cadmus-core": "^8.1.0",
-    "@myrmidon/cadmus-api": "^8.1.0",
-    "@myrmidon/cadmus-ui": "^8.1.2",
-    "@myrmidon/cadmus-state": "^8.1.0"
+    "@angular/common": "^19.0.0",
+    "@angular/core": "^19.0.0",
+    "@myrmidon/ng-tools": "^4.0.0",
+    "@myrmidon/ng-mat-tools": "^4.0.0",
+    "@myrmidon/paged-data-browsers": "^3.0.0",
+    "@myrmidon/cadmus-core": "^9.0.0",
+    "@myrmidon/cadmus-api": "^9.0.0",
+    "@myrmidon/cadmus-ui": "^9.0.0",
+    "@myrmidon/cadmus-state": "^9.0.0"
   },
   "dependencies": {
     "tslib": "^2.5.0"
