@@ -3,8 +3,7 @@ import { Component } from '@angular/core';
 @Component({
   selector: 'cadmus-flags-editor-feature',
   templateUrl: './flags-editor-feature.component.html',
-  styleUrls: ['./flags-editor-feature.component.css']
+  styleUrls: ['./flags-editor-feature.component.css'],
+  standalone: false,
 })
-export class FlagsEditorFeatureComponent {
-
-}
+export class FlagsEditorFeatureComponent {}
